requirements-dev.txt
# Core ML & Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# NLP
nltk==3.8.1

# Configuration
pyyaml==6.0

# Model Serialization
joblib>=1.3.0

# API & Web Services
flask==3.0.0
flask-cors==4.0.0

# Development & Testing
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Jupyter & exploration
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.13.0
